pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contact_study"
version = "0.1.0"
description = "Systematic study of contact model fidelity for sampling-based MPC in dexterous manipulation"
requires-python = ">=3.10"

dependencies = [
    "mujoco>=3.1",
    "warp-lang>=1.4",
    "numpy>=1.26",
    "matplotlib>=3.8",
    "scipy>=1.12",
]

[project.optional-dependencies]
dev = ["pytest", "black", "ruff"]

[tool.setuptools.packages.find]
where = ["."]
include = ["contact_study*"]

[tool.ruff]
line-length = 100
